Move the Tour into its own component
